add auth domain service
- decorator @require that checks if the scope in the user's token matches the role required for the operation
- added admin role protection to team create/delete

+++ src/tests/test_views.py
@@ -1,6 +1,8 @@
 import pytest
+from django.conf import settings
 
 from beheeromgeving.models import Team
+from tests.utils import build_jwt_token
 
 
 def test_health(api_client):
@@ -38,7 +40,8 @@ class TestViews:
         assert response.status_code == 200
         assert response.data["acronym"] == "DADI"
 
-    def test_teams_create(self, api_client):
+    def test_teams_create_by_admin(self, api_client):
+        token = build_jwt_token([settings.ADMIN_ROLE_NAME])
         response = api_client.post(
             "/teams",
             data={
@@ -50,10 +53,31 @@ class TestViews:
                 "contact_email": "benk@amsterdam.nl",
                 "scope": "scope_benk",
             },
+            HTTP_AUTHORIZATION=f"Bearer {token}",
         )
         assert response.status_code == 201
         result = api_client.get(f"/teams/{response.data}")
         assert result.data["acronym"] == "BENK"
+
+    def test_teams_create_unauthorized(self, api_client):
+        token = build_jwt_token(["some_unauthorized_scope"])
+        response = api_client.post(
+            "/teams",
+            data={
+                "name": "Basis- en Kernregistratie",
+                "description": "Beschrijving",
+                "acronym": "BENK",
+                "po_name": "Iemand",
+                "po_email": "iemand@amsterdam.nl",
+                "contact_email": "benk@amsterdam.nl",
+                "scope": "scope_benk",
+            },
+            HTTP_AUTHORIZATION=f"Bearer {token}",
+        )
+        assert response.status_code == 401
+        # No team created
+        result = api_client.get("/teams")
+        assert len(result.data) == 0
 
     def test_teams_update(self, api_client, orm_team):
         response = api_client.patch(
@@ -67,10 +91,17 @@ class TestViews:
         orm_team.refresh_from_db()
         assert orm_team.po_name == "Iemand Anders"
 
-    def test_teams_delete(self, api_client, orm_team):
-        response = api_client.delete(f"/teams/{orm_team.id}")
+    def test_teams_delete_by_admin(self, api_client, orm_team):
+        token = build_jwt_token([settings.ADMIN_ROLE_NAME])
+        response = api_client.delete(f"/teams/{orm_team.id}", HTTP_AUTHORIZATION=f"Bearer {token}")
         assert response.status_code == 204
         assert Team.objects.count() == 0
+
+    def test_teams_delete_unauthorized(self, api_client, orm_team):
+        token = build_jwt_token(["some_unauthorized_scope"])
+        response = api_client.delete(f"/teams/{orm_team.id}", HTTP_AUTHORIZATION=f"Bearer {token}")
+        assert response.status_code == 401
+        assert Team.objects.count() == 1
 
     def test_products_list(self, api_client, orm_product):
         response = api_client.get("/products")
